requirements.txt
packaging==24.1
pillow==11.0.0
PyMuPDF==1.24.11
pypdf==5.0.1
pytesseract==0.3.13
